+++ CHANGELOG.md
@@ -7,4 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added the option `break-build`, for the user to specify the job exit
+  status
+
+### Added
+
 - Started the project